pyproject.toml
# Email Agent - Python Project Configuration
#
# Dependencies and tooling for email-agent project.
# See PLAN.md §12 for authoritative specification.

[project]
name = "email-agent"
version = "0.0.1"
description = "Desktop AI agent that reads Gmail unread emails, categorizes them using Ollama LLM, and creates draft replies for important emails."
readme = "docs/README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
authors = [
    {name = "Email Agent Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "ollama>=0.5.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",
    "beautifulsoup4>=4.12.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
email-agent = "email_agent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/email_agent"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/docs",
]

# -----------------------------------------------------------------------
# Ruff Configuration
# -----------------------------------------------------------------------
# Linter and formatter for Python code quality.
# See: https://docs.astral.sh/ruff/

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "B",     # flake8-bugbear
    "C4",    # flake8-comprehensions
    "UP",    # pyupgrade
    "SIM",   # flake8-simplify
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E501",  # Line length handled by formatter
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# -----------------------------------------------------------------------
# MyPy Configuration
# -----------------------------------------------------------------------
# Static type checker for Python.
# See: https://mypy.readthedocs.io/

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

# Per-module overrides for test files (less strict)
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true

# Also ignore missing imports for the local email_agent package
[[tool.mypy.overrides]]
module = "email_agent.*"
ignore_missing_imports = true

# Ignore missing stubs for third-party packages
[[tool.mypy.overrides]]
module = "google.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "googleapiclient.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "google_auth_oauthlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "oauthlib.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "ollama.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bs4.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

# -----------------------------------------------------------------------
# Pytest Configuration
# -----------------------------------------------------------------------
# Test runner configuration.
# See: https://docs.pytest.org/

[tool.pytest]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.pytest_asyncio]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/email_agent"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.venv/*",
]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
